+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2026-01-19 - 2025-01-20
+### Features
+- Support ignore isolator for specific class using attribute `[Isolator(" ")]`.
+- Support copy properties in `BaseType` before and after execute methods. (Fix: #4)
+- Support dynamic or in-memory assembly by disabling isolator to avoid infinite loops. (Fix: #2)
+### Fixes
+- Fix issue with methods inside abstract class. (Fix: #3)
+### Updates
+- Add `GetFirstInstructionAfterBaseConstructor` to inject after base constructor. (Fix: #3)
+- Ignore isolator in abstract classes without `[Isolator]` attribute. (Fix: #3)
+### Tests
+- Add `AssemblyLoadContextTests` with `MemoryAssemblyLoadContext` tests.
+
 ## [1.0.0] / 2025-12-29 - 2025-01-09
 ### Features
 - Support attribute `[Isolator]` to mark classes for isolation.
@@ -43,4 +56,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Configuration` to use `LoadAtModuleInit` and force `Attach` to create default `AssemblyLoadContext` at module init.
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.0]: ../../compare/1.0.0...1.1.0
 [1.0.0]: ../../compare/1.0.0